pyproject.toml
[tool.poetry]
name = "upscale-shader-py"
version = "0.1.0"
description = ""
authors = ["Xuno"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyopengl = "^3.1.7"
numpy = "^2.2.1"
pillow = "^11.0.0"
glfw = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
